+++ R/manytrade-agreements.R
@@ -19,13 +19,13 @@
 #' \item{TREND: }{A dataset with 729 observations and the following
 #' 7 variables: manyID, Title, Beg, Signature, Force, treatyID, trendID.}
 #' \item{HUGGO: }{A dataset with 1391 observations and the following
-#' 15 variables: manyID, Title, Beg, Signature, Force, treatyID, TreatyText,
-#' url, MainText, AppendixText, AnnexText, End, Checked_HUGGO, Confirmed_HUGGO,
-#' Parties.}
+#' 10 variables: manyID, Title, Beg, Signature, Force, End, Parties,
+#' url, treatyID, TreatyTextStatus.}
 #' \item{AIGGO: }{A dataset with 1391 observations and the following
-#' 8 variables: manyID, Title, Beg, Signature, Force, accessionC, accessionP,
-#' treatyID.}
+#' 8 variables: manyID, Title, Beg, Signature, Force, accessionC,
+#' accessionP, treatyID.}
 #' }
+
 #'
 #' @details
 #' ``` {r, echo = FALSE, warning = FALSE}