+++ data-raw/agreements/HUGGO/prepare-HUGGO.R
@@ -262,12 +262,127 @@ HUGGO$End <- NA
 # Add Parties variable
 HUGGO$Parties <- ""
 
+# Add No_source variable
+#No_source_HUGGO code 1 when manual search online found no text
+# When blank it means that either a) the treaty has been checked and a source has been found or b) the treaty has not yet been checked
+
+HUGGO$No_source <- NA
+
+#Add Citation variable
+#Citation_HUGGO provides the citation when the source is from an academic journal, e.g jstor.com 
+#to help track down if URL breaks in the future
+HUGGO$Citation <- ""
+
+#Add MetaData_confirmed
+#MetaData_confirmed: 1 means that metadata (Signature, Force, End, Parties) has been checked and confirmed; 2 means that metadata has been checked and confirmed, but a variable could not be confirmed and is left blank
+#Variables with values can be considered confirmed, if left blank unconfirmed (almost always end date)
+HUGGO$MetaData_confirmed <- NA
+
+
 # Checked_HUGGO and Confirmed_HUGGO variables to track progress on manually correcting entries
 # Checked_HUGGO: code 1 when the entire row's observations have been verified and updated
 # Confirmed_HUGGO: list variables for which the observation could be verified and confirmed.
 # Eg. List 'Signature' in `Confirmed_HUGGO` if the Signature date was found and verified in the treaty text or in a manual online search.
 HUGGO$Checked_HUGGO <- NA
 HUGGO$Confirmed_HUGGO <- NA
+
+# Merging HUGGO_verified.csv with HUGGO 
+
+# 1) load HUGGO_verified.csv file & remove non-metadata columns
+HUGGO_verified <- read.csv2(file.choose())
+
+HUGGO_verified <- dplyr::select(HUGGO_verified, -Checked_HUGGO, -No_source, -MetaData_confirmed,)
+
+# 2) add updated 'Beg' column in HUGGO_verified using coalesce()
+HUGGO_verified <- dplyr::mutate(HUGGO_verified, Beg = dplyr::coalesce(Signature, "N/A"))
+
+# 3) change order of columns and name of RowNumbers variable
+HUGGO_verified <- dplyr::select(HUGGO_verified, manyID,Title,Beg,Signature,Force,treatyID,url,End,Parties, HUGGO_RowNumber)
+HUGGO_verified <- dplyr::rename(HUGGO_verified, RowNumbers = HUGGO_RowNumber )
+
+# 4) load agreements$HUGGO and add columns End + Parties + RowNumbers
+
+HUGGO <- agreements$HUGGO
+HUGGO$End <- NA
+HUGGO$Parties <- ""
+HUGGO <- HUGGO %>% mutate(RowNumbers = row_number())
+
+# 5) Add HUGGO TreatyText column to HUGGO_verified
+
+row_numbers <- HUGGO_verified$RowNumbers
+treaty_texts <- HUGGO$TreatyText[row_numbers]
+HUGGO_verified <- cbind(HUGGO_verified, TreatyText = treaty_texts)
+
+# 6) Reorder columns in HUGGO_verified  to match HUGGO
+
+HUGGO_verified <- HUGGO_verified[, c("manyID", "Title", "Beg", "Signature", "Force", "treatyID", "TreatyText", "url", "End", "Parties", "RowNumbers")]
+
+# 7) Join the two dataframes based on "RowNumbers"
+merged_df <- full_join(HUGGO, HUGGO_verified, by = "RowNumbers")
+
+# 8) remove variables that should not change
+merged_df <- dplyr::select(merged_df, -TreatyText.y, -Title.y, -treatyID.y, -RowNumbers)
+
+# 9) replace metadata for HUGGO_verified observations
+merged_df$Beg.x <- ifelse(!is.na(merged_df$manyID.y), merged_df$Beg.y, merged_df$Beg.x)
+merged_df$Signature.x <- ifelse(!is.na(merged_df$manyID.y), merged_df$Signature.y, merged_df$Signature.x)
+merged_df$Force.x <- ifelse(!is.na(merged_df$manyID.y), merged_df$Force.y, merged_df$Force.x)
+merged_df$url.x <- ifelse(!is.na(merged_df$manyID.y), merged_df$url.y, merged_df$url.x)
+merged_df$End.x <- ifelse(!is.na(merged_df$manyID.y), merged_df$End.y, merged_df$End.x)
+merged_df$Parties.x <- ifelse(!is.na(merged_df$manyID.y), merged_df$Parties.y, merged_df$Parties.x)
+
+# 10) remove .y variables and remove .x ending to column names
+merged_df <- dplyr::select(merged_df, -manyID.y, -Beg.y, -Signature.y, -Force.y, -url.y, -End.y, -Parties.y)
+
+merged_df <- dplyr::rename(merged_df, manyID = `1`,
+         Title = `2`,
+         Beg = `3`,
+         Signature = `4`,
+         Force = `5`,
+         treatyID = `6`,
+         TreatyText = `7`,
+         url = `8`,
+         End = `9`,
+         Parties = `10`)
+
+# 11) Apply messydates::as_messydate to Beg, Signature, Force, End columns
+merged_df <- dplyr::mutate(merged_df, Beg = as_messydate(Beg),
+         Signature = as_messydate(Signature),
+         Force = as_messydate(Force),
+         End = as_messydate(End))
+
+# 12) arrange merged_df by Beg Date and push to HUGGO
+
+merged_df <- dplyr::arrange(merged_df,Beg)
+
+# Remove TreatyText column, Add new TreatyTextStatus column
+# TreatyTextStatus: 0 = No text collected, 1 = Either Raw or formatted text collected
+HUGGO <- dplyr::select(HUGGO, -TreatyText)
+HUGGO$TreatyTextStatus <- 0
+
+#Get list of confirmed treaty texts collected in HUGGO_verified
+#Sorting TreatyTexts column, No_source = 0, otherwise = 1
+
+HUGGO_verified <- read.csv2(file.choose())
+HUGGO_verified$TreatyTextStatus <- 0
+HUGGO_verified$TreatyTextStatus[is.na(HUGGO_verified$No_source)] <- 1
+
+#integrate TreatyTexts column into corresponding rows
+
+HUGGO_verified <- dplyr::select(HUGGO_verified, manyID, url, Title, TreatyTextStatus)
+
+merged_df <- HUGGO %>%
+  dplyr::left_join(HUGGO_verified, by = c("manyID", "Title")) %>%
+  dplyr::mutate(TreatyTextStatus = if_else(is.na(TreatyTextStatus.y), TreatyTextStatus.x, TreatyTextStatus.y)) %>%
+  dplyr::select(-TreatyTextStatus.x, -TreatyTextStatus.y)
+
+#remove url.y columns and rename url.x as url, order by metadata
+merged_df <- dplyr::select(merged_df, -url.y)
+merged_df <- dplyr::rename(merged_df, url = url.x)
+merged_df <- dplyr::select(merged_df, manyID, Title, Beg, Signature, Force, End, Parties, url, treatyID, TreatyTextStatus)
+
+# push merged_df to HUGGO
+HUGGO <- merged_df
 
 # manypkgs includes several functions that should help cleaning
 # and standardising your data.