+++ NEWS.md
@@ -1,3 +1,11 @@
+# manytrade 0.2.1
+
+2023-03-02
+## Data
+
+* Closed issue Check Signature and Force dates, and add End dates #34 by verifying MetaData(Signature, Force, End, Parties, Url) for trade agreements in HUGGO dataset.
+* Added treaty texts for the agreements in HUGGO dataset.
+
 # manytrade 0.2.0
 
 2022-10-17